separate postbox/stash from blockhut
