+++ src/main/java/com/minecolonies/api/blocks/ModBlocks.java
@@ -49,12 +49,10 @@ public final class ModBlocks
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutStoneSmeltery;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutCrusher;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutSifter;
-    public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockPostBox;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutFlorist;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutEnchanter;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutUniversity;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutHospital;
-    public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockStash;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutSchool;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutGlassblower;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutDyer;
@@ -73,6 +71,9 @@ public final class ModBlocks
     //public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockLargeQuarry;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutAlchemist;
     public static AbstractBlockHut<? extends AbstractBlockHut<?>> blockHutKitchen;
+
+    public static AbstractBaseBlockHut<? extends AbstractBaseBlockHut<?>> blockStash;
+    public static AbstractBaseBlockHut<? extends AbstractBaseBlockHut<?>> blockPostBox;
 
     /**
      * Utility blocks.
@@ -120,9 +121,9 @@ public final class ModBlocks
     }
 
     @NotNull
-    public static AbstractBlockHut<?>[] getHuts()
+    public static AbstractBaseBlockHut<?>[] getHuts()
     {
-        return new AbstractBlockHut[] {
+        return new AbstractBaseBlockHut[] {
           blockHutTownHall,
           blockHutHome,
           blockHutTavern,
