+++ src/main/java/com/minecolonies/api/colony/buildings/registry/BuildingEntry.java
@@ -1,6 +1,7 @@
 package com.minecolonies.api.colony.buildings.registry;
 
 import com.google.common.collect.ImmutableMap;
+import com.minecolonies.api.blocks.AbstractBaseBlockHut;
 import com.minecolonies.api.blocks.AbstractBlockHut;
 import com.minecolonies.api.colony.IColony;
 import com.minecolonies.api.colony.IColonyView;
@@ -27,7 +28,7 @@ import java.util.function.Supplier;
 @SuppressWarnings("PMD.MissingStaticMethodInNonInstantiatableClass") //Use the builder to create one.
 public class BuildingEntry
 {
-    private final AbstractBlockHut<?> buildingBlock;
+    private final AbstractBaseBlockHut<?> buildingBlock;
 
     private final BiFunction<IColony, BlockPos, IBuilding> buildingProducer;
     private final ResourceLocation registryName;
@@ -39,7 +40,7 @@ public class BuildingEntry
      */
     public static final class Builder
     {
-        private AbstractBlockHut<?>                                        buildingBlock;
+        private AbstractBaseBlockHut<?>                                    buildingBlock;
         private BiFunction<IColony, BlockPos, IBuilding>                   buildingProducer;
         private Supplier<BiFunction<IColonyView, BlockPos, IBuildingView>> buildingViewProducer;
         private List<ModuleProducer>                                       buildingModuleProducers = new ArrayList<>();
@@ -51,7 +52,7 @@ public class BuildingEntry
          * @param buildingBlock The block.
          * @return The builder.
          */
-        public Builder setBuildingBlock(final AbstractBlockHut<?> buildingBlock)
+        public Builder setBuildingBlock(final AbstractBaseBlockHut<?> buildingBlock)
         {
             this.buildingBlock = buildingBlock;
             return this;
@@ -127,7 +128,7 @@ public class BuildingEntry
 
     private final Supplier<BiFunction<IColonyView, BlockPos, IBuildingView>> buildingViewProducer;
 
-    public AbstractBlockHut<?> getBuildingBlock()
+    public AbstractBaseBlockHut<?> getBuildingBlock()
     {
         return buildingBlock;
     }
@@ -165,7 +166,7 @@ public class BuildingEntry
 
     private BuildingEntry(
       final ResourceLocation registryName,
-      final AbstractBlockHut<?> buildingBlock,
+      final AbstractBaseBlockHut<?> buildingBlock,
       final BiFunction<IColony, BlockPos, IBuilding> buildingProducer,
       final Supplier<BiFunction<IColonyView, BlockPos, IBuildingView>> buildingViewProducer,
       List<ModuleProducer> buildingModuleProducers)
