+++ src/main/java/com/minecolonies/core/blocks/huts/BlockStash.java
@@ -1,5 +1,6 @@
 package com.minecolonies.core.blocks.huts;
 
+import com.minecolonies.api.blocks.AbstractBaseBlockHut;
 import com.minecolonies.api.blocks.AbstractBlockHut;
 import com.minecolonies.api.blocks.interfaces.IRSComponentBlock;
 import com.minecolonies.api.colony.IColonyManager;
@@ -29,7 +30,7 @@ import org.jetbrains.annotations.Nullable;
 /**
  * Hut for the Stash. No different from {@link AbstractBlockHut}
  */
-public class BlockStash extends AbstractBlockHut<BlockStash> implements IRSComponentBlock
+public class BlockStash extends AbstractBaseBlockHut<BlockStash> implements IRSComponentBlock
 {
 
     private static final VoxelShape SHAPE_NORTH = Block.box(0.0D, 0.0D, 0.0D, 16.0D, 16.0D, 8.0D);
